set space note to have no size and big selected note improvements

+++ lib/src/providers/list_of_spacing_for_each_row.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
 class ListOfSpacingForEachRow extends ChangeNotifier {
-  List<int> _rowSpacingList = [26];
+  List<double> _rowSpacingList = [26];
 
-  List<int> get rowSpacingList => _rowSpacingList;
+  List<double> get rowSpacingList => _rowSpacingList;
 
-  void updateRowSpacingList(List<int> rowSpacingList) {
+  void updateRowSpacingList(List<double> rowSpacingList) {
     _rowSpacingList = rowSpacingList;
     notifyListeners(); // Notify listeners about the change
   }
